moved all extension management utility to core
