+++ src/azure-cli-core/azure/cli/core/commands/__init__.py
@@ -27,7 +27,7 @@ from azure.cli.core.commands.constants import (
     CLI_POSITIONAL_PARAM_KWARGS, CONFIRM_PARAM_NAME)
 from azure.cli.core.commands.parameters import (
     AzArgumentContext, patch_arg_make_required, patch_arg_make_optional)
-from azure.cli.core.extension import get_extension
+from azure.cli.core.extensions import get_extension
 from azure.cli.core.util import get_command_type_kwarg, read_file_content, get_arg_list, poller_classes
 import azure.cli.core.telemetry as telemetry
