+++ src/azure-cli-core/azure/cli/core/commands/client_factory.py
@@ -10,7 +10,7 @@ from knack.util import CLIError
 
 from azure.cli.core import __version__ as core_version
 import azure.cli.core._debug as _debug
-from azure.cli.core.extension import EXTENSIONS_MOD_PREFIX
+from azure.cli.core.extensions import EXTENSIONS_MOD_PREFIX
 from azure.cli.core.profiles._shared import get_client_class, SDKProfile
 from azure.cli.core.profiles import ResourceType, CustomResourceType, get_api_version, get_sdk
