+++ src/azure-cli-core/azure/cli/core/extensions/_homebrew_patch.py
