+++ src/azure-cli-core/azure/cli/core/extensions/tests/latest/data/myfirstcliextension-0.0.4+dev-py2.py3-none-any.whl
