+++ src/azure-cli-core/azure/cli/core/extensions/tests/latest/data/notanextension.txt
