+++ src/command_modules/azure-cli-extension/azure/cli/command_modules/extension/_completers.py
@@ -4,9 +4,9 @@
 # --------------------------------------------------------------------------------------------
 
 from azure.cli.core.decorators import Completer
-from azure.cli.core.extension import get_extension_names
+from azure.cli.core.extensions import get_extension_names
 
-from azure.cli.command_modules.extension.custom import get_index_extensions
+from azure.cli.core.extensions.operations import get_index_extensions
 
 
 @Completer
